4-10-2018: Comit que corrige problemas de redondeo y elimina validación de de mínimo un rol y mínimo un item

+++ syncfusion_payc/syncfusion_payc/Controllers/DETALLE_FACTURA_ITEMController.cs
@@ -272,7 +272,7 @@ namespace syncfusion_payc.Controllers
                     connection.Close();
                 }
                 FACTURAS table1 = db.FACTURAS.Single(o => o.COD_FACTURA == param.value.COD_FACTURA);
-                table1.VALOR_SIN_IMPUESTOS = float.Parse(valor_factura);
+                table1.VALOR_SIN_IMPUESTOS = Decimal.Parse(valor_factura);
                 db.SaveChanges();
             }
             catch (Exception ex)
