+++ syncfusion_payc/syncfusion_payc/Models/CONTRATO_COLABORADOR_COMERCIAL.cs
@@ -34,13 +34,13 @@ public partial class CONTRATO_COLABORADOR_COMERCIAL
 
     public Nullable<long> MESES { get; set; }
 
-    public Nullable<float> VALOR_MENSUAL_SIN_PRESTACIONES { get; set; }
+    public Nullable<decimal> VALOR_MENSUAL_SIN_PRESTACIONES { get; set; }
 
-    public Nullable<float> VALOR_MENSUAL_PRESTACIONES { get; set; }
+    public Nullable<decimal> VALOR_MENSUAL_PRESTACIONES { get; set; }
 
-    public Nullable<float> VALOR_MENSUAL_PRESTACIONES_MULTIPLICADOR { get; set; }
+    public Nullable<decimal> VALOR_MENSUAL_PRESTACIONES_MULTIPLICADOR { get; set; }
 
-    public Nullable<float> DEDICACION { get; set; }
+    public Nullable<decimal> DEDICACION { get; set; }
 
     public string OBSERVACIONES { get; set; }
