+++ syncfusion_payc/syncfusion_payc/Models/FLUJO_INGRESOS_ROL.cs
@@ -28,11 +28,11 @@ public partial class FLUJO_INGRESOS_ROL
 
     public string ETAPA { get; set; }
 
-    public Nullable<float> VALOR_SIN_PRESTACIONES { get; set; }
+    public Nullable<decimal> VALOR_SIN_PRESTACIONES { get; set; }
 
-    public Nullable<float> VALOR_CON_PRESTACIONES { get; set; }
+    public Nullable<decimal> VALOR_CON_PRESTACIONES { get; set; }
 
-    public Nullable<float> VALOR_FACTOR_MULTIPLICADOR { get; set; }
+    public Nullable<decimal> VALOR_FACTOR_MULTIPLICADOR { get; set; }
 
     public string ESTADO { get; set; }
