+++ syncfusion_payc/syncfusion_payc/Models/GENERACION_FLUJOS_ROL3.cs
@@ -26,11 +26,11 @@ public partial class GENERACION_FLUJOS_ROL3
 
     public string ETAPA { get; set; }
 
-    public Nullable<float> VALOR_SIN_PRESTACIONES { get; set; }
+    public Nullable<decimal> VALOR_SIN_PRESTACIONES { get; set; }
 
-    public Nullable<float> VALOR_CON_PRESTACIONES { get; set; }
+    public Nullable<decimal> VALOR_CON_PRESTACIONES { get; set; }
 
-    public Nullable<float> VALOR_FACTOR_MULTIPLICADOR { get; set; }
+    public Nullable<decimal> VALOR_FACTOR_MULTIPLICADOR { get; set; }
 
 }
