+++ syncfusion_payc/syncfusion_payc/Models/ORDENES_SERVICIO_ROLES_COLABORADOR.cs
@@ -22,11 +22,11 @@ public partial class ORDENES_SERVICIO_ROLES_COLABORADOR
 
     public string ETAPA { get; set; }
 
-    public Nullable<float> VALOR_SIN_PRESTACIONES { get; set; }
+    public Nullable<decimal> VALOR_SIN_PRESTACIONES { get; set; }
 
-    public Nullable<float> VALOR_CON_PRESTACIONES { get; set; }
+    public Nullable<decimal> VALOR_CON_PRESTACIONES { get; set; }
 
-    public Nullable<float> VALOR_FACTOR_MULTIPLICADOR { get; set; }
+    public Nullable<decimal> VALOR_FACTOR_MULTIPLICADOR { get; set; }
 
     public string DESCRIPCION { get; set; }
