+++ syncfusion_payc/syncfusion_payc/Models/VISTA_FACTURAS_FECHA_PERIODO_FACTURACION.cs
@@ -30,9 +30,9 @@ public partial class VISTA_FACTURAS_FECHA_PERIODO_FACTURACION
 
     public Nullable<long> COD_ESTADO_FACTURA { get; set; }
 
-    public Nullable<float> VALOR_SIN_IMPUESTOS { get; set; }
+    public Nullable<decimal> VALOR_SIN_IMPUESTOS { get; set; }
 
-    public Nullable<float> VALOR_CON_IMPUESTOS { get; set; }
+    public Nullable<decimal> VALOR_CON_IMPUESTOS { get; set; }
 
     public string RANKING { get; set; }
